ON-16780: Remove additional tsc_hz calibration
It was found with certain warmup settings (mainly using time based
rather than iteration based limits) the recalibration of the tsc_hz
could be calculated completely differently. In a lot of testing it was
found that the minor calibration was almost always correct so do longer
calibration first and don't do the recalibration.